+++ mud-contracts/world/mud.config.ts
@@ -378,10 +378,10 @@ export default defineWorld({
      *************************/
     SmartTurretConfigTable: {
       schema: {
-        smartTurretId: "uint256",
+        smartObjectId: "uint256",
         systemId: "ResourceId",
       },
-      key: ["smartTurretId"],
+      key: ["smartObjectId"],
     },
 
     /*************************
@@ -389,11 +389,11 @@ export default defineWorld({
      *************************/
     SmartGateConfigTable: {
       schema: {
-        smartGateId: "uint256",
+        smartObjectId: "uint256",
         systemId: "ResourceId",
         maxDistance: "uint256",
       },
-      key: ["smartGateId"],
+      key: ["smartObjectId"],
     },
 
     SmartGateLinkTable: {