+++ mud-contracts/world/src/codegen/tables/SmartGateConfigTable.sol
@@ -42,7 +42,7 @@ library SmartGateConfigTable {
    */
   function getKeyNames() internal pure returns (string[] memory keyNames) {
     keyNames = new string[](1);
-    keyNames[0] = "smartGateId";
+    keyNames[0] = "smartObjectId";
   }
 
   /**
@@ -72,9 +72,9 @@ library SmartGateConfigTable {
   /**
    * @notice Get systemId.
    */
-  function getSystemId(uint256 smartGateId) internal view returns (ResourceId systemId) {
+  function getSystemId(uint256 smartObjectId) internal view returns (ResourceId systemId) {
     bytes32[] memory _keyTuple = new bytes32[](1);
-    _keyTuple[0] = bytes32(uint256(smartGateId));
+    _keyTuple[0] = bytes32(uint256(smartObjectId));
 
     bytes32 _blob = StoreSwitch.getStaticField(_tableId, _keyTuple, 0, _fieldLayout);
     return ResourceId.wrap(bytes32(_blob));
@@ -83,9 +83,9 @@ library SmartGateConfigTable {
   /**
    * @notice Get systemId.
    */
-  function _getSystemId(uint256 smartGateId) internal view returns (ResourceId systemId) {
+  function _getSystemId(uint256 smartObjectId) internal view returns (ResourceId systemId) {
     bytes32[] memory _keyTuple = new bytes32[](1);
-    _keyTuple[0] = bytes32(uint256(smartGateId));
+    _keyTuple[0] = bytes32(uint256(smartObjectId));
 
     bytes32 _blob = StoreCore.getStaticField(_tableId, _keyTuple, 0, _fieldLayout);
     return ResourceId.wrap(bytes32(_blob));
@@ -94,9 +94,9 @@ library SmartGateConfigTable {
   /**
    * @notice Set systemId.
    */
-  function setSystemId(uint256 smartGateId, ResourceId systemId) internal {
+  function setSystemId(uint256 smartObjectId, ResourceId systemId) internal {
     bytes32[] memory _keyTuple = new bytes32[](1);
-    _keyTuple[0] = bytes32(uint256(smartGateId));
+    _keyTuple[0] = bytes32(uint256(smartObjectId));
 
     StoreSwitch.setStaticField(_tableId, _keyTuple, 0, abi.encodePacked(ResourceId.unwrap(systemId)), _fieldLayout);
   }
@@ -104,9 +104,9 @@ library SmartGateConfigTable {
   /**
    * @notice Set systemId.
    */
-  function _setSystemId(uint256 smartGateId, ResourceId systemId) internal {
+  function _setSystemId(uint256 smartObjectId, ResourceId systemId) internal {
     bytes32[] memory _keyTuple = new bytes32[](1);
-    _keyTuple[0] = bytes32(uint256(smartGateId));
+    _keyTuple[0] = bytes32(uint256(smartObjectId));
 
     StoreCore.setStaticField(_tableId, _keyTuple, 0, abi.encodePacked(ResourceId.unwrap(systemId)), _fieldLayout);
   }
@@ -114,9 +114,9 @@ library SmartGateConfigTable {
   /**
    * @notice Get maxDistance.
    */
-  function getMaxDistance(uint256 smartGateId) internal view returns (uint256 maxDistance) {
+  function getMaxDistance(uint256 smartObjectId) internal view returns (uint256 maxDistance) {
     bytes32[] memory _keyTuple = new bytes32[](1);
-    _keyTuple[0] = bytes32(uint256(smartGateId));
+    _keyTuple[0] = bytes32(uint256(smartObjectId));
 
     bytes32 _blob = StoreSwitch.getStaticField(_tableId, _keyTuple, 1, _fieldLayout);
     return (uint256(bytes32(_blob)));
@@ -125,9 +125,9 @@ library SmartGateConfigTable {
   /**
    * @notice Get maxDistance.
    */
-  function _getMaxDistance(uint256 smartGateId) internal view returns (uint256 maxDistance) {
+  function _getMaxDistance(uint256 smartObjectId) internal view returns (uint256 maxDistance) {
     bytes32[] memory _keyTuple = new bytes32[](1);
-    _keyTuple[0] = bytes32(uint256(smartGateId));
+    _keyTuple[0] = bytes32(uint256(smartObjectId));
 
     bytes32 _blob = StoreCore.getStaticField(_tableId, _keyTuple, 1, _fieldLayout);
     return (uint256(bytes32(_blob)));
@@ -136,9 +136,9 @@ library SmartGateConfigTable {
   /**
    * @notice Set maxDistance.
    */
-  function setMaxDistance(uint256 smartGateId, uint256 maxDistance) internal {
+  function setMaxDistance(uint256 smartObjectId, uint256 maxDistance) internal {
     bytes32[] memory _keyTuple = new bytes32[](1);
-    _keyTuple[0] = bytes32(uint256(smartGateId));
+    _keyTuple[0] = bytes32(uint256(smartObjectId));
 
     StoreSwitch.setStaticField(_tableId, _keyTuple, 1, abi.encodePacked((maxDistance)), _fieldLayout);
   }
@@ -146,9 +146,9 @@ library SmartGateConfigTable {
   /**
    * @notice Set maxDistance.
    */
-  function _setMaxDistance(uint256 smartGateId, uint256 maxDistance) internal {
+  function _setMaxDistance(uint256 smartObjectId, uint256 maxDistance) internal {
     bytes32[] memory _keyTuple = new bytes32[](1);
-    _keyTuple[0] = bytes32(uint256(smartGateId));
+    _keyTuple[0] = bytes32(uint256(smartObjectId));
 
     StoreCore.setStaticField(_tableId, _keyTuple, 1, abi.encodePacked((maxDistance)), _fieldLayout);
   }
@@ -156,9 +156,9 @@ library SmartGateConfigTable {
   /**
    * @notice Get the full data.
    */
-  function get(uint256 smartGateId) internal view returns (SmartGateConfigTableData memory _table) {
+  function get(uint256 smartObjectId) internal view returns (SmartGateConfigTableData memory _table) {
     bytes32[] memory _keyTuple = new bytes32[](1);
-    _keyTuple[0] = bytes32(uint256(smartGateId));
+    _keyTuple[0] = bytes32(uint256(smartObjectId));
 
     (bytes memory _staticData, EncodedLengths _encodedLengths, bytes memory _dynamicData) = StoreSwitch.getRecord(
       _tableId,
@@ -171,9 +171,9 @@ library SmartGateConfigTable {
   /**
    * @notice Get the full data.
    */
-  function _get(uint256 smartGateId) internal view returns (SmartGateConfigTableData memory _table) {
+  function _get(uint256 smartObjectId) internal view returns (SmartGateConfigTableData memory _table) {
     bytes32[] memory _keyTuple = new bytes32[](1);
-    _keyTuple[0] = bytes32(uint256(smartGateId));
+    _keyTuple[0] = bytes32(uint256(smartObjectId));
 
     (bytes memory _staticData, EncodedLengths _encodedLengths, bytes memory _dynamicData) = StoreCore.getRecord(
       _tableId,
@@ -186,14 +186,14 @@ library SmartGateConfigTable {
   /**
    * @notice Set the full data using individual values.
    */
-  function set(uint256 smartGateId, ResourceId systemId, uint256 maxDistance) internal {
+  function set(uint256 smartObjectId, ResourceId systemId, uint256 maxDistance) internal {
     bytes memory _staticData = encodeStatic(systemId, maxDistance);
 
     EncodedLengths _encodedLengths;
     bytes memory _dynamicData;
 
     bytes32[] memory _keyTuple = new bytes32[](1);
-    _keyTuple[0] = bytes32(uint256(smartGateId));
+    _keyTuple[0] = bytes32(uint256(smartObjectId));
 
     StoreSwitch.setRecord(_tableId, _keyTuple, _staticData, _encodedLengths, _dynamicData);
   }
@@ -201,14 +201,14 @@ library SmartGateConfigTable {
   /**
    * @notice Set the full data using individual values.
    */
-  function _set(uint256 smartGateId, ResourceId systemId, uint256 maxDistance) internal {
+  function _set(uint256 smartObjectId, ResourceId systemId, uint256 maxDistance) internal {
     bytes memory _staticData = encodeStatic(systemId, maxDistance);
 
     EncodedLengths _encodedLengths;
     bytes memory _dynamicData;
 
     bytes32[] memory _keyTuple = new bytes32[](1);
-    _keyTuple[0] = bytes32(uint256(smartGateId));
+    _keyTuple[0] = bytes32(uint256(smartObjectId));
 
     StoreCore.setRecord(_tableId, _keyTuple, _staticData, _encodedLengths, _dynamicData, _fieldLayout);
   }
@@ -216,14 +216,14 @@ library SmartGateConfigTable {
   /**
    * @notice Set the full data using the data struct.
    */
-  function set(uint256 smartGateId, SmartGateConfigTableData memory _table) internal {
+  function set(uint256 smartObjectId, SmartGateConfigTableData memory _table) internal {
     bytes memory _staticData = encodeStatic(_table.systemId, _table.maxDistance);
 
     EncodedLengths _encodedLengths;
     bytes memory _dynamicData;
 
     bytes32[] memory _keyTuple = new bytes32[](1);
-    _keyTuple[0] = bytes32(uint256(smartGateId));
+    _keyTuple[0] = bytes32(uint256(smartObjectId));
 
     StoreSwitch.setRecord(_tableId, _keyTuple, _staticData, _encodedLengths, _dynamicData);
   }
@@ -231,14 +231,14 @@ library SmartGateConfigTable {
   /**
    * @notice Set the full data using the data struct.
    */
-  function _set(uint256 smartGateId, SmartGateConfigTableData memory _table) internal {
+  function _set(uint256 smartObjectId, SmartGateConfigTableData memory _table) internal {
     bytes memory _staticData = encodeStatic(_table.systemId, _table.maxDistance);
 
     EncodedLengths _encodedLengths;
     bytes memory _dynamicData;
 
     bytes32[] memory _keyTuple = new bytes32[](1);
-    _keyTuple[0] = bytes32(uint256(smartGateId));
+    _keyTuple[0] = bytes32(uint256(smartObjectId));
 
     StoreCore.setRecord(_tableId, _keyTuple, _staticData, _encodedLengths, _dynamicData, _fieldLayout);
   }
@@ -269,9 +269,9 @@ library SmartGateConfigTable {
   /**
    * @notice Delete all data for given keys.
    */
-  function deleteRecord(uint256 smartGateId) internal {
+  function deleteRecord(uint256 smartObjectId) internal {
     bytes32[] memory _keyTuple = new bytes32[](1);
-    _keyTuple[0] = bytes32(uint256(smartGateId));
+    _keyTuple[0] = bytes32(uint256(smartObjectId));
 
     StoreSwitch.deleteRecord(_tableId, _keyTuple);
   }
@@ -279,9 +279,9 @@ library SmartGateConfigTable {
   /**
    * @notice Delete all data for given keys.
    */
-  function _deleteRecord(uint256 smartGateId) internal {
+  function _deleteRecord(uint256 smartObjectId) internal {
     bytes32[] memory _keyTuple = new bytes32[](1);
-    _keyTuple[0] = bytes32(uint256(smartGateId));
+    _keyTuple[0] = bytes32(uint256(smartObjectId));
 
     StoreCore.deleteRecord(_tableId, _keyTuple, _fieldLayout);
   }
@@ -315,9 +315,9 @@ library SmartGateConfigTable {
   /**
    * @notice Encode keys as a bytes32 array using this table's field layout.
    */
-  function encodeKeyTuple(uint256 smartGateId) internal pure returns (bytes32[] memory) {
+  function encodeKeyTuple(uint256 smartObjectId) internal pure returns (bytes32[] memory) {
     bytes32[] memory _keyTuple = new bytes32[](1);
-    _keyTuple[0] = bytes32(uint256(smartGateId));
+    _keyTuple[0] = bytes32(uint256(smartObjectId));
 
     return _keyTuple;
   }