+++ mud-contracts/world/src/codegen/world/ISmartTurretSystem.sol
@@ -15,10 +15,10 @@ import { TargetPriority, Turret, SmartTurretTarget } from "../../modules/smart-t
  */
 interface ISmartTurretSystem {
   error SmartTurret_UndefinedClassId();
-  error SmartTurret_NotConfigured(uint256 smartTurretId);
+  error SmartTurret_NotConfigured(uint256 smartObjectId);
 
   function eveworld__createAndAnchorSmartTurret(
-    uint256 smartTurretId,
+    uint256 smartObjectId,
     EntityRecordData memory entityRecordData,
     SmartObjectData memory smartObjectData,
     WorldPosition memory worldPosition,
@@ -27,10 +27,10 @@ interface ISmartTurretSystem {
     uint256 fuelMaxCapacity
   ) external;
 
-  function eveworld__configureSmartTurret(uint256 smartTurretId, ResourceId systemId) external;
+  function eveworld__configureSmartTurret(uint256 smartObjectId, ResourceId systemId) external;
 
   function eveworld__inProximity(
-    uint256 smartTurretId,
+    uint256 smartObjectId,
     uint256 turretOwnerCharacterId,
     TargetPriority[] memory priorityQueue,
     Turret memory turret,
@@ -38,7 +38,7 @@ interface ISmartTurretSystem {
   ) external returns (TargetPriority[] memory updatedPriorityQueue);
 
   function eveworld__aggression(
-    uint256 smartTurretId,
+    uint256 smartObjectId,
     uint256 turretOwnerCharacterId,
     TargetPriority[] memory priorityQueue,
     Turret memory turret,